Use new no-entity agent setup flag
Previously if the rackspace_monitoring_service's create_entity attribute
was false, the call to setup the agent would completely hang the Chef
run while the agent prompted for input.  Version 2.2.10 of the agent
adds a `--no-entity` flag [1] to avoid this.  This uses that flag if
create_entity is false.  Older versions of the agent will ignore the
flag.

  [1]: https://github.com/virgo-agent-toolkit/rackspace-monitoring-agent/pull/910